+++ alooma/alooma_exceptions.py
@@ -1,0 +1,2 @@
+class FailedToCreateInputException(Exception):
+    pass
